requirements.txt
torch~=2.3.1
torchvision~=0.18.1
pygame~=2.6.1
pillow~=10.4.0
tqdm~=4.66.4
matplotlib~=3.9.1
